requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
Flask_MySQLdb==2.0.0
matplotlib==3.7.2
nltk==3.8.1
numpy==1.24.3
pytz==2023.3.post1
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.5.0
seaborn==0.13.2
torch==2.0.1
